improve mediagroup send method & ugoira to mp4 add trycatch in db_inital fix format mssing => error

+++ handlers/pixiv/tools.js
@@ -106,7 +106,7 @@ async function thumb_to_all(illust, try_time = 0) {
  * @param {*} force ignore exist file 
  * @returns 
  */
-async function ugoira_to_mp4(id, retry_time = 0, force = false) {
+async function ugoira_to_mp4(id, force = false, retry_time = 0) {
     if (fs.existsSync(`./tmp/mp4_1/${id}.mp4`) && !force) {
         return `${config.pixiv.ugoiraurl}/${id}.mp4`
     }
@@ -115,7 +115,7 @@ async function ugoira_to_mp4(id, retry_time = 0, force = false) {
     }
     if (ugoira_mp4_queue_list.length > 4 || ugoira_mp4_queue_list.includes(id)) {
         await sleep(1000)
-        return await ugoira_to_mp4(id, false)
+        return await ugoira_to_mp4(id, force, retry_time)
     }
     ugoira_mp4_queue_list.push(id)
     try {
@@ -135,16 +135,9 @@ async function ugoira_to_mp4(id, retry_time = 0, force = false) {
         fs.writeFileSync(`./tmp/timecode/${id}`, frame)
         // 下载
         await download_file(ud.originalSrc, id)
-        // Windows 自己补全这些软件并且自己改路径之类的 不做兼容
-        // 解压没有现成好用的轮子
-        // 所以干脆直接 exec 了 以后有好办法再改咯
         // force 为强制更新
-        if (fs.existsSync(`./tmp/mp4_1/${id}.mp4`)) {
-            if (!force) {
-                fs.unlinkSync(`./tmp/mp4_1/${id}.mp4`)
-            } else {
-                return `${config.pixiv.ugoiraurl}/mp4_1/${id}.mp4`
-            }
+        if (fs.existsSync(`./tmp/mp4_1/${id}.mp4`) && force) {
+            fs.unlinkSync(`./tmp/mp4_1/${id}.mp4`)
         }
         if (fs.existsSync(`./tmp/ugoira/${id}`)) {
             fs.rmdirSync(`./tmp/ugoira/${id}`, {
@@ -154,27 +147,40 @@ async function ugoira_to_mp4(id, retry_time = 0, force = false) {
         if (fs.existsSync(`./tmp/mp4_0/${id}.mp4`)) {
             fs.unlinkSync(`./tmp/mp4_0/${id}.mp4`)
         }
+        // windows:
+        // choco install ffmpeg unzip
+
+        // 解压没有现成好用的轮子
+        // 所以干脆直接 exec 了 以后有好办法再改咯
         await exec(`unzip -n './tmp/file/${id}.zip' -d './tmp/ugoira/${id}'`)
         // copy last frame
+        // see this issue https://github.com/my-telegram-bots/Pixiv_bot/issues/1
         fs.copyFileSync(`./tmp/ugoira/${id}/${(ud.frames.length - 1).toString().padStart(6, 0)}.jpg`, `./tmp/ugoira/${id}/${(ud.frames.length).toString().padStart(6, 0)}.jpg`)
         // jpg -> mp4 (no fps metadata)
         await exec(`ffmpeg -y -i ./tmp/ugoira/${id}/%6d.jpg -c:v libx264 -vf "format=yuv420p,scale=trunc(iw/2)*2:trunc(ih/2)*2" ./tmp/mp4_0/${id}.mp4`, { timeout: 240 * 1000 })
-        // add time metadata via mp4fpsmod
+        // add fps metadata via mp4fpsmod
         await exec(`mp4fpsmod -o ./tmp/mp4_1/${id}.mp4 -t ./tmp/timecode/${id} ./tmp/mp4_0/${id}.mp4`, { timeout: 240 * 1000 })
         ugoira_mp4_queue_list.splice(ugoira_mp4_queue_list.indexOf(id), 1)
+        // pre handle mp4 to gif
+        ugoira_to_gif(id, 'large')
+        ugoira_to_gif(id, 'medium')
+        ugoira_to_gif(id, 'small')
         return `${config.pixiv.ugoiraurl}/${id}.mp4`
     } catch (error) {
         honsole.warn(error)
         ugoira_mp4_queue_list.splice(ugoira_mp4_queue_list.indexOf(id), 1)
-        return await ugoira_to_mp4(id, retry_time + 1, force)
+        return await ugoira_to_mp4(id, force, retry_time + 1)
     }
 }
 /**
  * ugoira mp4 to gif
  * ~~ why not apng to gif ? ~~ -> lazy
- * @param {*} id 
+ * @param {number} id 
+ * @param {string} quality
+ * @param {number} real_width
+ * @param {number} real_height
  */
-async function ugoira_to_gif(id, quality = 'large', real_width = 0, real_height = 0, retry_time = 0, force = false) {
+async function ugoira_to_gif(id, quality = 'large', real_width = 0, real_height = 0, force = false, retry_time = 0) {
     let height = 0
     let width = 0
     if (!['large', 'medium', 'small'].includes(quality)) {
@@ -185,7 +191,7 @@ async function ugoira_to_gif(id, quality = 'large', real_width = 0, real_height 
     }
     if (ugoira_gif_queue_list.length > 4 || ugoira_gif_queue_list.includes(id)) {
         await sleep(1000)
-        return await ugoira_to_gif(id, quality, real_width, real_height, retry_time, force)
+        return await ugoira_to_gif(id, quality, real_width, real_height, force, retry_time)
     }
     ugoira_gif_queue_list.push(id)
     await ugoira_to_mp4(id)
@@ -223,11 +229,8 @@ async function ugoira_to_gif(id, quality = 'large', real_width = 0, real_height 
         console.warn(error)
         ugoira_gif_queue_list.splice(ugoira_gif_queue_list.indexOf(id), 1)
         await sleep(500)
-        return await ugoira_to_gif(id, quality, real_width, real_height, retry_time + 1, force)
+        return await ugoira_to_gif(id, quality, real_width, real_height, force, retry_time + 1)
     }
-    ugoira_to_gif(id,'large')
-    ugoira_to_gif(id,'medium')
-    ugoira_to_gif(id,'small')
     return `${config.pixiv.ugoiraurl.replace('mp4_1', 'gif')}/${id}-${quality}.gif`
 }
 /**
